Change how we conditionally compile access to externs
Now, everything that deals with low-level intrinsicts are wrapped within `#[cfg(not(test))]` block so they are compiled only outside of tests.

What this achieves is we ensure people are able to write tests properly for their substreams and don't hit compilation problem due to their target architecture being wrong.